removed muiicons from package.json, renamed Logsfile folder to Logs, renamed functions, organized state, removed unused imports
