+++ ui/src/pages/Logs/Logstyles.css
